requirements.txt
langchain==0.3.27
langchain-core==0.3.76
python-dotenv==1.1.1
beautifulsoup4==4.13.5
fastapi==0.116.2
html5lib==1.1
jinja2==3.1.6
langchain-astradb==0.6.1
langchain-google-genai==2.1.12
langchain-groq==0.3.8
lxml==6.0.1
python-multipart==0.0.20
selenium==4.32.0
streamlit==1.49.1
undetected-chromedriver==3.5.5
uvicorn==0.35.0
structlog==25.4.0
langgraph==0.6.7
ragas==0.3.4

-e .